+++ src/contexts/DataProvider.jsx
@@ -1,6 +1,6 @@
 import { createContext, useContext, useState } from 'react';
-import { useAuth } from './AuthProvider';
 import { getWordsList } from '../utils/firebase';
+import { useAuth } from './AuthProvider';
 
 export const DataContext = createContext();
 export const useData = () => useContext(DataContext);
@@ -11,15 +11,18 @@ export const DataProvider = ({ children }) => {
 
   const getWordsCollection = async () => {
     if (currentUser) {
-      getWordsList(currentUser.uid)
-         .then((resp) => setWordsCollection(resp));
+      return getWordsList(currentUser.uid).then((resp) =>
+        setWordsCollection(resp)
+      );
     }
+
+    return [];
   };
 
   const value = {
     wordsCollection,
     getWordsCollection,
-  }
+  };
 
   return <DataContext.Provider value={value}>{children}</DataContext.Provider>;
 };
