+++ src/views/userSections/UserSection.jsx
@@ -1,0 +1,12 @@
+import { SectionWrapper } from '../../components/MainSectionWrapper';
+import { ActionsSection } from './actionsSection/ActionsSection';
+import { WordSection } from './wordsSection/WordsSection';
+
+export const UserSection = () => {
+  return (
+    <SectionWrapper>
+      <ActionsSection />
+      <WordSection />
+    </SectionWrapper>
+  );
+};
