+++ src/views/userSections/actionsSection/ActionsNavigation.jsx
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import { Button } from '../../../../components/Button';
+import { Button } from '../../../components/Button';
 
 export const ActionsNavigation = ({ onAddWorld }) => {
   return (
