+++ src/views/userSections/actionsSection/actionsSection.style.scss
@@ -1,4 +1,4 @@
-@use '../../../../variables' as *;
+@use '../../../variables' as *;
 
 .actions {
    width: 100vw;
