+++ src/views/userSections/wordsSection/wordsSection.style.scss
@@ -1,4 +1,4 @@
-@use '../../../../variables' as *;
+@use '../../../variables' as *;
 
 .words {
    width: 93vw;
